1.3.1: Hit tables and more
* ENH cleanup Buff_manager - fix warnings, use named temporary vars ("name", "buff"), unify iteration schemes
BF increment_stat_gains() was using an invalid iterator after erase()
BF increment_hit_gains() wasn't erasing hit_effects_{mh,oh} correctly
ENH increment_over_time_buffs() was using updating min_interval for each change
BF modify_arpen_stacks() was adding arpen at full OH stacks
ENH hit tables are now recomputed on expertise changes, too
BF Deep Wounds is now correctly updating its tick damage on each crit

ENH added "integration tests", simply featuring reference results (test_fury, test_2h); these should probably use more interesting gear (hit_effects, etc.)

* ENH Weapon_sim is now directly based on Weapon, and calculates average_damage on construction
ENH simplified compute_hit_tables(), it's using pre-calculated values

* BF (modelling) all yellows now use two-roll tables
BF there's now split yellow tables for main- and off-hand (e.g. when wielding different weapon types)
BF removed overpower's arbitrary -3% crit
ENH prepared new style Hit_tables, and Hit_outcome
ENH updated "integration test" results

* ENH use new-style hit tables
ENH don't use vector<Hit_outcome> in whirlwind() and swing_weapon()
BF this also fixes ww oh hits sometimes using the oh_queued hit table

* BF dodged OH whites granted slightly too much rage (they weren't using the offhand penalty); added new results to "integration tests"

* ENH simplify rage_generation(), now that "rageDamage" is available directly in hitOutcome

* BF/ENH the rampage cast didn't grant +50 AP (in game, it starts at 1/5 stacks, effectively), so it only granted +200 AP when fully stacked, and losing rampage resulted in a net -50 AP; rampage_active has been replaced by rampage_stacks > 0

* ENH/BF split manage_flurry_rampage() into 3 more specialized functions; cleave and whirlwind now add rampage stacks for each individual hit, not just once

* BF mace spec was only proccing of white swings or next-melee abilities

* BF heroic strike and cleave are now unqueued before hit_effects() (and therefore possible extra attacks) again

* BF incomplete refactoring

* BF glancing damage % was wrongly displayed

+++ statistics/include/Distribution.hpp
@@ -27,7 +27,7 @@ public:
     double variance_;
     double max_;
     double min_;
-    int64_t n_samples_;
+    int n_samples_;
 };
 
 #endif // WOW_SIMULATOR_DISTRIBUTION_HPP
